:heavy_plus_sign: :green_heart: Worked on improving my code
The code overall was very confusing and hard to read for me who isn't that advanced enough to be comfortable with directly implementing what I had learnt. Therefore, I decided to try rewriting the models in my own words, allowing me to have a better understanding of what was happening. There is still an issue with the members displaying but the members can now be stored as a data inside each club that has been made.